+++ src/microservices/payment/gateways/vnpay.gateway.example.ts
@@ -11,7 +11,7 @@
  */
 
 import { Injectable, Logger } from '@nestjs/common';
-import { IPaymentGateway, PaymentGatewayResponse, PaymentWebhookResult } from './payment-gateway.interface';
+import { IPaymentGateway, PaymentGatewayResponse, PaymentWebhookResult } from '../interfaces/payment-gateway.interface';
 import { Payment } from 'src/shared/entities/payment/payment.entity';
 import { Booking } from 'src/shared/entities/booking/booking.entity';
 // import VNPay from 'vnpay'; // Uncomment khi install package