+++ src/microservices/payment/interfaces/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Payment Service Interfaces
+ * All interfaces used by Payment Service are exported from here
+ */
+
+export * from './payment-gateway.interface';
+