+++ src/microservices/search/interfaces/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Search Service Interfaces
+ * All interfaces used by Search Service are exported from here
+ */
+
+export * from './flight-result.interface';
+